Fix for latest_api references
The "latest_api" link in the top-level docs folder redirects
you to the Asterisk_21_Documentation section but doesn't change
the nav bar because "latest_api" doesn't actually appear in
the menu.  To make that work, the link has been renamed to
"Latest_API" and is now shown in the nav bar.