astxml2markdown.xslt: Parse new CEP information.
Gosub and Goto were not displaying the parameters context, extension,
and priority in the correct order. This is due to the fact that, while
context and extension are optional parameters, they came before
priority, which is required, but there was no way to specify reverse
order for preceding optional parameters. The stylesheet now looks for
'documentationtype' in the parameter. If it is 'dialplan_context' or
'dialplan_extension', it will be ignored. If it is 'dialplan_priority',
the following will be injected:

[[context,]extension,]priority

This is treated as one expression, so cases that added brackets for
optional parameters also needed to be tweaked.

Fixes: #92