mkdocs.yml: Add missing 'site_url' keyword
mkdocs.yml was missing the 'site_url' keyword which was preventing
mkdocs from generating the sitemap.xml file.  It's been added with
the default of 'https://docs.asterisk.org' but can be overridden
in Makefile.inc for local deployment and testing.  README.md has
been updated with instructions.

Also removed a duplicate 'features' section in the yml and fixed
the BRANCHES example in the readme.